+++ extensions/badges/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Badges Changelog
 
+## [New Features] - 2024-05-06
+
+- Add `Create Social Badge` command
+- Add `Create Dynamic Badge` command
+- Add `Create Endpoint Badge` command
+- Polish extension icon with gradient overlay
+- Add Color Picker support
+
 ## [Initial Version] - 2024-04-30
 
 Concise, consistent, and legible badges with [shields.io](https://github.com/badges/shields).